Move OTEL tracer to a sub package
